Take `OwnedFd` instead of `RawFd` in `new_from_fd`
This is a breaking API change and requires Rust 1.63.

The function still needs to be `unsafe` since the fd is mmaped, but this
is more idiomatic and makes it clear that the file descriptor is owned
rather than borrowed. (Which should match the existing implementation).
Though maybe it should only borrow the FD?