Fix panic in key_get_utf8
At Zed, we see a "capacity overflow" from some users in this function.

I don't know under what circumstances exactly, but the documentation for
xkb_state_get_utf8 says it has a return value "similar" to snprintf,
which may return a negative number on unspecified error, or a length
larger than the buffer in the case of overflow.